Fix flicker with popup menu
When popup menu elements have additional infos associated those are displayed in a preview window. When the preview window is shown the cursor is briefly switched to the preview window and back to its original position thus triggering a BufLeave and corresponding BufEnter command. This can cause a function call, that may trigger a redraw causing a flicker while moving between popup menu result. Only executing those function when the popup menu is not visible fixes the issue.