Better Fold Text-Object
Before, the <kbd>esc</kbd> and `:normal` was unnecessary because it's already in the normal mode and will create a bell. The original `viz` will select the beginning marker if the foldmethod was `syntax` or `marker`. Now, it will select the real `inner` as long as the content starts in the next line of the beginning marker.